feat: Complete basic L2 distance endpoint for person-to-person

+++ app/backend/classes/Location.py
@@ -7,13 +7,12 @@ from api.config.databases import sql_db as db
 #     latitude: float
 #     longitude: float
 #     name: str
-    
 
 class Location(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(100))
-    latitude = db.Column(db.Integer)
-    longitude = db.Column(db.Integer)
+    latitude = db.Column(db.Float)
+    longitude = db.Column(db.Float)
     
     def __repr__(self):
         return f"Location {self.name} at {(self.longitude, self.latitude)} with id {self.id}"
